solved the issue of mixing the files content of server.js file ans swagger.js file
